+++ src/main/java/com/yuuko/core/scheduler/jobs/ThirtySecondlyJob.java
@@ -1,12 +1,14 @@
 package com.yuuko.core.scheduler.jobs;
 
 import com.yuuko.core.scheduler.Job;
+import com.yuuko.core.scheduler.tasks.RenewShardTask;
 import com.yuuko.core.scheduler.tasks.UpdateMetricsPingTask;
 
 import java.util.concurrent.TimeUnit;
 
 public class ThirtySecondlyJob extends Job {
     private final UpdateMetricsPingTask updateMetricsPingTask = new UpdateMetricsPingTask();
+    private final RenewShardTask renewShardTask = new RenewShardTask();
 
     public ThirtySecondlyJob() {
         super(0, 30, TimeUnit.SECONDS);
@@ -14,6 +16,6 @@ public class ThirtySecondlyJob extends Job {
 
     @Override
     public void run() {
-        handleTask(updateMetricsPingTask);
+        handleTask(updateMetricsPingTask, renewShardTask);
     }
 }